Add low-level support for the XSPEC rgsxsrc convolution model
This model was part of XSPEC version 11 but removed in initial versions
of XSPEC 12, being added back in XSPEC 12.8.1. Somehow I missed this
so there was no support for it in the XSPEC module. I have also
added commented-out support for the convolution models added in XSPEC
12.9.1, since we do not have a way of tailoring the XSPEC module
to the XSPEC version it is being compiled againts.

Adding support for rgsxsrc requires adding new template code, since
it is the only convolution model that uses the F77 XSPEC model
interface (the others use the C interface). I did the minimum job
here (using copy and paste), rather than refactoring the templates
as there is now a lot of repeated logic.

As the convoultion models are only provided at a very-low level, there
are now tests of them, so none is added here for rgsxsrc. Note that
the idea of this interface is to allow us to experiment with what
is needed in Sherpa to properly support XSPEC convolution-style
models.