Inital attempt at moving AREASCAL handling into instrument-model classes
To support Poisson statistics, and to better match XSPEC, the handling
of the AREASCAL values has been moved more into the Astronomy instrument
models (specifically, the ARF/RMF/RSP models which have an associated
PHA data set). This almost treats the AREASCAL value as a "correction"
to the exposure time.

With this approach, the hope is that the DataPHA object no longer has
to worry about the AREASCAL values, but it unfortunately doesn't work
out: background subtraction, the error contribution of any backround
component, and calculating the "y axis" values used to create plots
(which includes 'correcting' the model values) has to be AREASCAL
aware.

This commit is a move towards this, but it has known issues: there is
a new test added in this commit which has been marked xfail as I
haven't had time to work out what is going on; the handling of backgrounds
and errors in the DataPHA is littered with "is this sensible" comments,
and from a brief look with actual data it does not appear to be doing
the right thing. It has been committed as the code has been worked on
for long enough that I don't want to lose what has been done so far,
and to provide enough of an example of the proposed approach.